+++ src/main/java/app/visa/repository/DemandeurRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.visa.entity.Demandeur;
 
-public interface DemandeurRepository extends JpaRepository<Demandeur, Long> {
+public interface DemandeurRepository extends JpaRepository<Demandeur, Integer> {
 }
